producer/requirements.txt
kafka-python==2.0.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1